Renamed/Added/Fixed Bybit REST examples
