+++ examples/bybit/previous_version/bybit_inverse_perpetual_v2.rs
@@ -16,14 +16,14 @@ async fn main() {
 
     let open_interest: serde_json::Value = client.get(
         "/v2/public/open-interest",
-        Some(&[("symbol", "BTCUSDT"), ("period", "1d"), ("limit", "3")]),
+        Some(&[("symbol", "BTCUSD"), ("period", "1d"), ("limit", "3")]),
         [BybitOption::HttpAuth(BybitHttpAuth::None)],
     ).await.expect("failed to cancel orders");
     println!("Open interest:\n{}", open_interest);
 
     let positions: serde_json::Value = client.get(
         "/v2/private/position/list",
-        Some(&[("symbol", "BTCUSDT")]),
+        Some(&[("symbol", "BTCUSD")]),
         [BybitOption::HttpAuth(BybitHttpAuth::BelowV3)],
     ).await.expect("failed to get positions");
     println!("Positions:\n{}", positions);
