+++ examples/bybit/previous_version/bybit_spot_v1.rs
