+++ examples/bybit/previous_version/bybit_usdc_contract_websocket_public.rs
