+++ examples/bybit/previous_version/bybit_usdc_option_v1.rs
@@ -1,6 +1,5 @@
 use std::env;
 use log::LevelFilter;
-use serde_json::json;
 use crypto_botters::{Client, bybit::{BybitOption, BybitHttpAuth}};
 
 #[tokio::main]
@@ -15,25 +14,23 @@ async fn main() {
     client.update_default_option(BybitOption::Secret(secret));
 
     // private POST
-    let cancel_all: serde_json::Value = client.post(
-        "/perpetual/usdc/openapi/private/v1/cancel-all",
-        Some(json!({"symbol": "BTCPERP", "orderFilter": "Order"})),
-        [BybitOption::HttpAuth(BybitHttpAuth::V3AndAbove)],
+    let cancel_all: serde_json::Value = client.post_no_body(
+        "/option/usdc/openapi/private/v1/cancel-all",
+        [BybitOption::HttpAuth(BybitHttpAuth::UsdcContractV1)],
     ).await.expect("failed to cancel orders");
     println!("Cancel all result:\n{}", cancel_all);
 
-    // private POST
-    let open_orders: serde_json::Value = client.post(
-        "/option/usdc/openapi/private/v1/query-active-orders",
-        Some(json!({"category": "PERPETUAL"})),
-        [BybitOption::HttpAuth(BybitHttpAuth::V3AndAbove)],
+    // private GET
+    let open_orders: serde_json::Value = client.get_no_query(
+        "/option/usdc/openapi/private/v1/trade/query-active-orders",
+        [BybitOption::HttpAuth(BybitHttpAuth::UsdcContractV1)],
     ).await.expect("failed to get orders");
     println!("Open orders:\n{}", open_orders);
 
     // public GET
     let symbol_info: serde_json::Value = client.get(
-        "/perpetual/usdc/openapi/public/v1/tick",
-        Some(&[("symbol", "BTCPERP")]),
+        "/option/usdc/openapi/public/v1/tick",
+        Some(&[("symbol", "BTC-5JAN23-18500-C")]),
         [BybitOption::Default],
     ).await.expect("failed to get symbol info");
     println!("Symbol info:\n{}", symbol_info);
