+++ examples/bybit/previous_version/bybit_usdt_contract_websocket_public.rs
