+++ examples/bybit/v3/bybit_spot_v3_websocket_public.rs
