+++ src/exchanges/bybit.rs
@@ -96,12 +96,18 @@ pub enum BybitWebSocketUrl {
 /// Represents the auth type.
 #[derive(Debug, Eq, PartialEq, Copy, Clone)]
 pub enum BybitHttpAuth {
-    /// "Previous Version" APIs except for [Spot V1](https://bybit-exchange.github.io/docs-legacy/spot/v1/#t-introduction)
-    BelowV3,
     /// [Spot V1](https://bybit-exchange.github.io/docs-legacy/spot/v1/#t-introduction)
     SpotV1,
+    /// "Previous Version" APIs except for [Spot V1](https://bybit-exchange.github.io/docs-legacy/spot/v1/#t-introduction),
+    /// [USDC Option](https://bybit-exchange.github.io/docs-legacy/usdc/option/#t-introduction), and
+    /// [USDC Perpetual](https://bybit-exchange.github.io/docs-legacy/usdc/perpetual/#t-introduction)
+    BelowV3,
+    /// [USDC Option](https://bybit-exchange.github.io/docs-legacy/usdc/option/#t-introduction) and
+    /// [USDC Perpetual](https://bybit-exchange.github.io/docs-legacy/usdc/perpetual/#t-introduction)
+    UsdcContractV1,
     /// [V3](https://bybit-exchange.github.io/docs/v3/intro) and [V5](https://bybit-exchange.github.io/docs/v5/intro)
     V3AndAbove,
+    /// No authentication (for public APIs)
     None,
 }
 
@@ -160,9 +166,10 @@ where
         let hmac = Hmac::<Sha256>::new_from_slice(secret.as_bytes()).unwrap(); // hmac accepts key of any length
 
         match self.options.http_auth {
-            BybitHttpAuth::BelowV3 => Self::v1_auth(builder, request_body, key, timestamp, hmac, false, self.options.recv_window),
             BybitHttpAuth::SpotV1 => Self::v1_auth(builder, request_body, key, timestamp, hmac, true, self.options.recv_window),
-            BybitHttpAuth::V3AndAbove => Self::v3_auth(builder, request_body, key, timestamp, hmac, self.options.recv_window),
+            BybitHttpAuth::BelowV3 => Self::v1_auth(builder, request_body, key, timestamp, hmac, false, self.options.recv_window),
+            BybitHttpAuth::UsdcContractV1 => Self::v3_auth(builder, request_body, key, timestamp, hmac, true, self.options.recv_window),
+            BybitHttpAuth::V3AndAbove => Self::v3_auth(builder, request_body, key, timestamp, hmac, false, self.options.recv_window),
             BybitHttpAuth::None => unreachable!(), // we've already handled this case
         }
     }
@@ -195,7 +202,7 @@ where
 
 impl<'a, R> BybitRequestHandler<'a, R> where R: DeserializeOwned {
     fn v1_auth<B>(builder: RequestBuilder, request_body: &Option<B>, key: &str, timestamp: u128, mut hmac: Hmac<Sha256>, spot: bool, window: Option<i32>)
-                  -> Result<Request, <BybitRequestHandler<'a, R> as RequestHandler<B>>::BuildError>
+        -> Result<Request, <BybitRequestHandler<'a, R> as RequestHandler<B>>::BuildError>
     where
         B: Serialize,
     {
@@ -299,8 +306,8 @@ impl<'a, R> BybitRequestHandler<'a, R> where R: DeserializeOwned {
         Ok(request)
     }
 
-    fn v3_auth<B>(mut builder: RequestBuilder, request_body: &Option<B>, key: &str, timestamp: u128, mut hmac: Hmac<Sha256>, window: Option<i32>)
-                  -> Result<Request, <BybitRequestHandler<'a, R> as RequestHandler<B>>::BuildError>
+    fn v3_auth<B>(mut builder: RequestBuilder, request_body: &Option<B>, key: &str, timestamp: u128, mut hmac: Hmac<Sha256>, version_header: bool, window: Option<i32>)
+        -> Result<Request, <BybitRequestHandler<'a, R> as RequestHandler<B>>::BuildError>
     where
         B: Serialize,
     {
@@ -338,6 +345,9 @@ impl<'a, R> BybitRequestHandler<'a, R> where R: DeserializeOwned {
         let signature = hex::encode(hmac.finalize().into_bytes());
 
         let headers = request.headers_mut();
+        if version_header {
+            headers.insert("X-BAPI-SIGN-TYPE", HeaderValue::from(2));
+        }
         headers.insert("X-BAPI-SIGN", HeaderValue::from_str(&signature).unwrap()); // hex digits are valid
         headers.insert("X-BAPI-API-KEY", HeaderValue::from_str(key).or(Err("invalid character in API key"))?);
         headers.insert("X-BAPI-TIMESTAMP", HeaderValue::from(timestamp as u64));
